refactor menu data from mock to a composable
