+++ src/mocks/layout.ts
@@ -1,74 +1,3 @@
-import { Link } from 'src/components/types';
-
-// config
-import { routesConf } from 'src/router/routes_conf';
-
-export const menuTop: Link[] = [
-  {
-    url: routesConf['home']['children']['fullPath'],
-    icon: 'svguse:icons/drawer_menu/icons.svg#lucide-home',
-    name: 'home',
-    title: 'home',
-  },
-  {
-    url: routesConf['routes']['children']['fullPath'],
-    icon: 'svguse:icons/drawer_menu/icons.svg#lucide-route',
-    name: 'routes',
-    title: 'routes',
-    disabled: true,
-  },
-  {
-    url: routesConf['results']['children']['fullPath'],
-    icon: 'svguse:icons/drawer_menu/icons.svg#chart-graph',
-    name: 'results',
-    title: 'results',
-    disabled: true,
-  },
-  // {
-  //   url: '',
-  //   icon: 'svguse:icons/drawer_menu/icons.svg#three-circles',
-  //   name: 'community',
-  //   title: 'community',
-  // },
-  // {
-  //   url: '',
-  //   icon: 'svguse:icons/drawer_menu/icons.svg#lucide-badge-percent',
-  //   name: 'prizes',
-  //   title: 'discounts',
-  // },
-  {
-    url: routesConf['coordinator']['children']['fullPath'],
-    icon: 'svguse:icons/drawer_menu/icons.svg#lucide-building',
-    name: 'coordinator',
-    title: 'coordinator',
-  },
-  {
-    url: routesConf['profile']['children']['fullPath'],
-    icon: 'svguse:icons/drawer_menu/icons.svg#ion-person-outline',
-    name: 'profile',
-    title: 'profile',
-  },
-];
-
-export const getMenuBottom = (url: string): Link[] => {
-  const menuBottom: Link[] = [
-    // {
-    //   url: '',
-    //   icon: 'svguse:icons/drawer_menu/icons.svg#email',
-    //   name: 'invite',
-    //   title: 'inviteFriends',
-    // },
-    {
-      url: '',
-      icon: 'svguse:icons/drawer_menu/icons.svg#lucide-gift',
-      name: 'donate',
-      title: 'donate',
-      href: url,
-    },
-  ];
-  return menuBottom;
-};
-
 const dummyText =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.';
